# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LocumLink Kenya API — two-sided marketplace for medical staffing
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /locums:
    get:
      operationId: listLocums
      tags: [locums]
      summary: List verified locum professionals
      parameters:
        - name: specialtyId
          in: query
          schema:
            type: integer
        - name: verificationStatus
          in: query
          schema:
            type: string
        - name: subCounty
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of locums
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocumList"
    post:
      operationId: createLocum
      tags: [locums]
      summary: Create a locum profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocumBody"
      responses:
        "201":
          description: Locum created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Locum"
...
  /locums/{id}:
    get:
      operationId: getLocum
      tags: [locums]
      summary: Get a locum by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Locum details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Locum"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLocum
      tags: [locums]
      summary: Update locum profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLocumBody"
      responses:
        "200":
          description: Updated locum
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Locum"
  /locums/{id}/availability:
    get:
      operationId: getLocumAvailability
      tags: [locums]
      summary: Get locum availability calendar
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: string
            description: "Format: YYYY-MM"
      responses:
        "200":
          description: Availability slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityList"
    post:
      operationId: setLocumAvailability
      tags: [locums]
      summary: Set availability slots
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetAvailabilityBody"
      responses:
        "200":
          description: Availability updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # SHIFTS
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payment history for the current user
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Payment history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentList"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/earnings-summary:
    get:
      operationId: getEarningsSummary
      tags: [payments]
      summary: Locum earnings summary with KRA-ready tax breakdown
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Earnings summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EarningsSummary"

  # RATINGS
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for the current user
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationList"
  /notifications/preferences:
    get:
      operationId: getNotificationPreferences
      tags: [notifications]
      summary: Get notification preferences
      responses:
        "200":
          description: Preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
    patch:
      operationId: updateNotificationPreferences
      tags: [notifications]
      summary: Update notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNotificationPreferencesBody"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ANALYTICS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # AUTH SCHEMAS
...
    Specialty:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        suggestedRateMin:
          type: integer
        suggestedRateMax:
          type: integer
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - suggestedRateMin
        - suggestedRateMax
...
    Locum:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        bio:
          type: string
        primarySpecialtyId:
          type: integer
        registrationNumber:
          type: string
        registrationBody:
          type: string
        yearsExperience:
          type: integer
        preferredRatePerShift:
          type: integer
        verificationStatus:
          type: string
        reliabilityScore:
          type: string
        totalShiftsCompleted:
          type: integer
        profilePhotoUrl:
          type: string
        isAvailableForUrgent:
          type: boolean
        subCounty:
          type: string
        county:
          type: string
        certExpiryDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        specialty:
          $ref: "#/components/schemas/Specialty"
      required:
        - id
        - firstName
        - lastName
        - registrationNumber
        - registrationBody
        - verificationStatus
    LocumList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Locum"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    CreateLocumBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        registrationNumber:
          type: string
        registrationBody:
          type: string
        primarySpecialtyId:
          type: integer
        yearsExperience:
          type: integer
        mpesaNumber:
          type: string
        bio:
          type: string
      required:
        - firstName
        - lastName
        - registrationNumber
        - registrationBody
        - primarySpecialtyId
        - yearsExperience
    UpdateLocumBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        bio:
          type: string
        primarySpecialtyId:
          type: integer
        yearsExperience:
          type: integer
        mpesaNumber:
          type: string
        preferredRatePerShift:
          type: integer
        isAvailableForUrgent:
          type: boolean
        subCounty:
          type: string
        profilePhotoUrl:
          type: string
        kraPin:
          type: string
    AvailabilitySlot:
      type: object
      properties:
        id:
          type: integer
        locumId:
          type: integer
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        isAvailable:
          type: boolean
        isRecurring:
          type: boolean
        recurringDayOfWeek:
          type: integer
      required:
        - id
        - locumId
        - date
        - isAvailable
    AvailabilityList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AvailabilitySlot"
      required:
        - data
    SetAvailabilityBody:
      type: object
      properties:
        slots:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              startTime:
                type: string
              endTime:
                type: string
              isAvailable:
                type: boolean
              isRecurring:
                type: boolean
              recurringDayOfWeek:
                type: integer
            required:
              - date
              - isAvailable
      required:
        - slots

    # SHIFT SCHEMAS
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        grossAmount:
          type: integer
        platformFee:
          type: integer
        locumPayout:
          type: integer
        paymentMethod:
          type: string
        status:
          type: string
        mpesaTransactionId:
          type: string
        invoiceUrl:
          type: string
        invoiceNumber:
          type: string
        paidAt:
          type: string
          format: date-time
        releasedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - bookingId
        - grossAmount
        - platformFee
        - locumPayout
        - paymentMethod
        - status
    PaymentList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    EarningsSummary:
      type: object
      properties:
        totalEarnings:
          type: integer
        totalShifts:
          type: integer
        averagePerShift:
          type: integer
        platformFeesDeducted:
          type: integer
        netEarnings:
          type: integer
        byMonth:
          type: array
          items:
            type: object
            properties:
              month:
                type: string
              earnings:
                type: integer
              shifts:
                type: integer
            required:
              - month
              - earnings
              - shifts
        bySpecialty:
          type: array
          items:
            type: object
            properties:
              specialtyName:
                type: string
              earnings:
                type: integer
              shifts:
                type: integer
            required:
              - specialtyName
              - earnings
              - shifts
        taxYear:
          type: integer
        kraReadyGrossIncome:
          type: integer
      required:
        - totalEarnings
        - totalShifts
        - netEarnings
        - kraReadyGrossIncome

    # RATING SCHEMAS
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        channel:
          type: string
        title:
          type: string
        content:
          type: string
        status:
          type: string
        sentAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - channel
        - title
        - content
        - status
    NotificationList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total:
          type: integer
        unreadCount:
          type: integer
      required:
        - data
        - total
        - unreadCount
    NotificationPreferences:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        newShiftMatch:
          type: boolean
        applicationUpdate:
          type: boolean
        bookingConfirmation:
          type: boolean
        paymentUpdate:
          type: boolean
        shiftReminder:
          type: boolean
        ratingReceived:
          type: boolean
        credentialExpiry:
          type: boolean
        preferWhatsapp:
          type: boolean
        preferSms:
          type: boolean
        preferEmail:
          type: boolean
      required:
        - newShiftMatch
        - applicationUpdate
        - bookingConfirmation
        - paymentUpdate
        - shiftReminder
        - ratingReceived
        - credentialExpiry
        - preferWhatsapp
        - preferSms
        - preferEmail
    UpdateNotificationPreferencesBody:
      type: object
      properties:
        newShiftMatch:
          type: boolean
        applicationUpdate:
          type: boolean
        bookingConfirmation:
          type: boolean
        paymentUpdate:
          type: boolean
        shiftReminder:
          type: boolean
        ratingReceived:
          type: boolean
        credentialExpiry:
          type: boolean
        preferWhatsapp:
          type: boolean
        preferSms:
          type: boolean
        preferEmail:
          type: boolean

    # ANALYTICS SCHEMAS
...